refactor: hoc을 만들어 인증 처리 로직 분리
