+++ src/pages/LoginPage.tsx
@@ -2,22 +2,9 @@ import styled from "@emotion/styled";
 import TheHeader from "@/components/layout/TheHeader";
 import LoginForm from "@/components/login/LoginForm";
 import KakaoLogo from "@/components/image/KakaoLogo";
-import { useUserInfo } from "@/contexts/UserInfoContext";
-import { useEffect } from "react";
-import { useNavigate } from "react-router";
-import { ROUTE_PATH } from "@/routes/paths";
+import withoutUser from "@/hoc/withoutUser";
 
 const LoginPage = () => {
-  const navigate = useNavigate();
-  const user = useUserInfo();
-
-  useEffect(() => {
-    if (user?.email) {
-      const redirectPath = new URLSearchParams(location.search).get("redirect");
-      navigate(redirectPath || ROUTE_PATH.HOME, { replace: true });
-    }
-  }, [user, navigate]);
-
   return (
     <>
       <TheHeader />
@@ -29,7 +16,7 @@ const LoginPage = () => {
   );
 };
 
-export default LoginPage;
+export default withoutUser(LoginPage);
 
 const Main = styled.main`
   display: flex;
